Developing example

+++ example/src/main/java/org/teavm/flavour/example/api/ProductDTO.java
@@ -15,8 +15,6 @@
  */
 package org.teavm.flavour.example.api;
 
-import java.math.BigDecimal;
-
 /**
  *
  * @author Alexey Andreev
@@ -25,5 +23,5 @@ public class ProductDTO {
     public int id;
     public String sku;
     public String name;
-    public BigDecimal unitPrice;
+    public double unitPrice;
 }
